third_party/WebKit/LayoutTests/platform/win/virtual/rootlayerscrolls/fast/scrolling/scrollbar-prevent-default-expected.txt

This tests that scrollbars always receive events even when there is an element underneath the scrollbar which swallows the event. In this case an event handler is added to the window which swallows the event but this may also happen with elements underneath of overlay scrollbars.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL receivedMousedownEvent should be true. Was false.
FAIL window.scrollY failed to change to 0 in 0.5 seconds.
PASS receivedMousedownEvent is true
FAIL subframe.scrollTop failed to change to 0 in 0.5 seconds.
PASS successfullyParsed is true

TEST COMPLETE
